requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
pytest==6.1.2
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
requests==2.25.0
urllib3==1.26.2
validators==0.18.2